整理日志名称
op_type	用途	contextpick_{task_type}	领取任务	页名return_{task_type}	退回任务	页名submit_{task_type}	提交任务	页名save_do_{task_type}	新任务保存	页名save_update_{task_type}	原任务保存	页名save_edit_{task_type}	任务修改保存	页名save_cmp_{task_type}	比对文本保存	页名publish_{task_type}	发布任务	2个任务: JX_165_7_12,JX_165_7_20submit_{task_type}	提交任务	页名submit_{task_type}	提交任务	页名get_users	取用户列表	角色管理取到 9 个用户login-no-user	登录的账号不存在	phone_or_emaillogin-fail	密码不对	phone_or_emailregister	注册	email, phone, namelogout	注销	change_user_profile	修改用户信息	张三: phone, namechange_role	修改用户角色	张三: 栏切分校对员reset_password	重置密码	张三delete_user	删除用户	张三change_password	自主修改密码	change_profile	修改个人信息	visit	打开页面	request.path

+++ controller/user/api.py
@@ -137,7 +137,7 @@ class RegisterApi(BaseHandler):
                 create_time=hlp.get_date_time()
             ))
             user['_id'] = r.inserted_id
-            self.add_op_log('register', context='%s: %s: %s' % (user.get('email'), user.get('phone'), user['name']))
+            self.add_op_log('register', context='%s, %s, %s' % (user.get('email'), user.get('phone'), user['name']))
         except DbError as e:
             return self.send_db_error(e)
 
@@ -178,7 +178,7 @@ class ChangeUserProfileApi(BaseHandler):
 
             r = self.db.user.update_one(dict(_id=user['_id']), {'$set': sets})
             if r.modified_count:
-                self.add_op_log('change_user_profile', context='%s: %s' % (user['_id'], ','.join(sets.keys())))
+                self.add_op_log('change_user_profile', context='%s: %s' % (old_user['name'], ','.join(sets.keys())))
 
             self.send_data_response(dict(info=sets))
 
@@ -199,11 +199,15 @@ class ChangeUserRoleApi(BaseHandler):
             return self.send_error_response(err)
 
         try:
+            old_user = self.db.user.find_one(dict(_id=user['_id']))
+            if not old_user:
+                return self.send_error_response(errors.no_user, id=user['_id'])
+
             user['roles'] = user.get('roles') or ''
             r = self.db.user.update_one(dict(_id=user['_id']), {'$set': dict(roles=user['roles'])})
             if not r.matched_count:
                 return self.send_error_response(errors.no_user)
-            self.add_op_log('change_role', context='%s: %s' % (user.get('_id'), user.get('roles')))
+            self.add_op_log('change_role', context='%s: %s' % (old_user['name'], user.get('roles')))
         except DbError as e:
             return self.send_db_error(e)
         self.send_data_response({'roles': user['roles']})
@@ -230,7 +234,7 @@ class ResetUserPasswordApi(BaseHandler):
 
             user = self.db.user.find_one(dict(_id=oid))
             self.remove_login_fails(self, user['_id'])
-            self.add_op_log('reset_password', context=': '.join(user))
+            self.add_op_log('reset_password', context=user['name'])
         except DbError as e:
             return self.send_db_error(e)
         self.send_data_response({'password': pwd})
@@ -256,12 +260,15 @@ class DeleteUserApi(BaseHandler):
             return self.send_error_response(err)
 
         try:
+            old_user = self.db.user.find_one(dict(_id=user['_id']))
+            if not old_user:
+                return self.send_error_response(errors.no_user, id=user['_id'])
             if user['_id'] == self.current_user['_id']:  # 判断删除的用户是否为自己
                 return self.send_error_response(errors.cannot_delete_self)
             r = self.db.user.delete_one(dict(_id=user['_id']))
             if r.deleted_count < 1:
                 return self.send_error_response(errors.no_user)
-            self.add_op_log('delete_user', context=': '.join(user))
+            self.add_op_log('delete_user', context=old_user['name'])
         except DbError as e:
             return self.send_db_error(e)
         self.send_data_response()
